+++ stellar_sdk/operation/allow_trust.py
@@ -1,9 +1,9 @@
 from .operation import Operation
+from .utils import check_ed25519_public_key, check_asset_code
 from ..asset import Asset
 from ..keypair import Keypair
-from ..xdr import Xdr
 from ..strkey import StrKey
-from .utils import check_ed25519_public_key, check_asset_code
+from ..xdr import xdr
 
 
 class AllowTrust(Operation):
@@ -38,52 +38,47 @@ class AllowTrust(Operation):
         self.authorize: bool = authorize
 
     @classmethod
-    def type_code(cls) -> int:
-        return Xdr.const.ALLOW_TRUST
+    def type_code(cls) -> xdr.OperationType:
+        return xdr.OperationType.ALLOW_TRUST
 
-    def _to_operation_body(self) -> Xdr.nullclass:
+    def _to_operation_body(self) -> xdr.OperationBody:
         Asset.check_if_asset_code_is_valid(self.asset_code)
         trustor = Keypair.from_public_key(self.trustor).xdr_account_id()
         length = len(self.asset_code)
-        # assert length <= 12
         pad_length = 4 - length if length <= 4 else 12 - length
-        # asset_code = self.asset_code + '\x00' * pad_length
-        # asset_code = bytearray(asset_code, encoding='utf-8')
         asset_code = bytearray(self.asset_code, "ascii") + b"\x00" * pad_length
-        asset = Xdr.nullclass()
         if len(asset_code) == 4:
-            asset.type = Xdr.const.ASSET_TYPE_CREDIT_ALPHANUM4
-            asset.assetCode4 = asset_code
+            asset_type = xdr.AssetType.ASSET_TYPE_CREDIT_ALPHANUM4
+            asset_code4 = xdr.AssetCode4(asset_code)
+            asset = xdr.AllowTrustOpAsset(type=asset_type, asset_code4=asset_code4)
         else:
-            asset.type = Xdr.const.ASSET_TYPE_CREDIT_ALPHANUM12
-            asset.assetCode12 = asset_code
-
-        allow_trust_op = Xdr.types.AllowTrustOp(trustor, asset, self.authorize)
-        body = Xdr.nullclass()
-        body.type = Xdr.const.ALLOW_TRUST
-        body.allowTrustOp = allow_trust_op
+            asset_type = xdr.AssetType.ASSET_TYPE_CREDIT_ALPHANUM12
+            asset_code12 = xdr.AssetCode12(asset_code)
+            asset = xdr.AllowTrustOpAsset(type=asset_type, asset_code12=asset_code12)
+        allow_trust_op = xdr.AllowTrustOp(trustor, asset, self.authorize)
+        body = xdr.OperationBody(type=self.type_code(), allow_trust_op=allow_trust_op)
         return body
 
     @classmethod
-    def from_xdr_object(cls, operation_xdr_object: Xdr.types.Operation) -> "AllowTrust":
+    def from_xdr_object(cls, operation_xdr_object: xdr.Operation) -> "AllowTrust":
         """Creates a :class:`AllowTrust` object from an XDR Operation
         object.
 
         """
         source = Operation.get_source_from_xdr_obj(operation_xdr_object)
         trustor = StrKey.encode_ed25519_public_key(
-            operation_xdr_object.body.allowTrustOp.trustor.ed25519
+            operation_xdr_object.body.allow_trust_op.trustor.account_id.ed25519.uint256
         )
-        authorize = operation_xdr_object.body.allowTrustOp.authorize
+        authorize = operation_xdr_object.body.allow_trust_op.authorize
 
-        asset_type = operation_xdr_object.body.allowTrustOp.asset.type
-        if asset_type == Xdr.const.ASSET_TYPE_CREDIT_ALPHANUM4:
+        asset_type = operation_xdr_object.body.allow_trust_op.asset.type
+        if asset_type == xdr.AssetType.ASSET_TYPE_CREDIT_ALPHANUM4:
             asset_code = (
-                operation_xdr_object.body.allowTrustOp.asset.assetCode4.decode()
+                operation_xdr_object.body.allow_trust_op.asset.asset_code4.asset_code4.decode()
             )
-        elif asset_type == Xdr.const.ASSET_TYPE_CREDIT_ALPHANUM12:
+        elif asset_type == xdr.AssetType.ASSET_TYPE_CREDIT_ALPHANUM12:
             asset_code = (
-                operation_xdr_object.body.allowTrustOp.asset.assetCode12.decode()
+                operation_xdr_object.body.allow_trust_op.asset.asset_code12.asset_code12.decode()
             )
         else:
             raise NotImplementedError(