+++ tests/test_memo.py
@@ -10,11 +10,12 @@ from stellar_sdk.memo import NoneMemo, Memo, TextMemo, IdMemo, HashMemo, ReturnH
 class TestMemo:
     def test_none_memo(self):
         memo = NoneMemo()
-        assert memo.to_xdr_object().to_xdr() == "AAAAAA=="
+        assert memo.to_xdr() == "AAAAAA=="
+        assert Memo.from_xdr("AAAAAA==") == memo
 
         base_memo = Memo.from_xdr_object(memo.to_xdr_object())
         assert isinstance(base_memo, NoneMemo)
-        assert base_memo.to_xdr_object().to_xdr() == "AAAAAA=="
+        assert base_memo.to_xdr() == "AAAAAA=="
 
     @pytest.mark.parametrize(
         "text, xdr",
@@ -26,11 +27,12 @@ class TestMemo:
     )
     def test_text_memo(self, text, xdr):
         memo = TextMemo(text)
-        assert memo.to_xdr_object().to_xdr() == xdr
+        assert memo.to_xdr() == xdr
+        assert Memo.from_xdr(xdr) == memo
 
         base_memo = Memo.from_xdr_object(memo.to_xdr_object())
         assert isinstance(base_memo, TextMemo)
-        assert base_memo.to_xdr_object().to_xdr() == xdr
+        assert base_memo.to_xdr() == xdr
 
     def test_text_memo_invalid_type_raise(self):
         invalid_value = 123
@@ -62,11 +64,12 @@ class TestMemo:
     )
     def test_id_memo(self, id, xdr):
         memo = IdMemo(id)
-        assert memo.to_xdr_object().to_xdr() == xdr
+        assert memo.to_xdr() == xdr
+        assert Memo.from_xdr(xdr) == memo
 
         base_memo = Memo.from_xdr_object(memo.to_xdr_object())
         assert isinstance(base_memo, IdMemo)
-        assert base_memo.to_xdr_object().to_xdr() == xdr
+        assert base_memo.to_xdr() == xdr
 
     @pytest.mark.parametrize("id", [-1, 2 ** 64])
     def test_id_memo_invalid_raise(self, id):
@@ -76,63 +79,62 @@ class TestMemo:
         ):
             IdMemo(id)
 
-    @pytest.mark.parametrize(
-        "hex, xdr",
-        [
-            (
-                "90d38f2d64949bb64f693ff231f9a3b3",
-                "AAAAA5DTjy1klJu2T2k/8jH5o7MAAAAAAAAAAAAAAAAAAAAA",
-            ),
-            (
-                "573c10b148fc4bc7db97540ce49da22930f4bcd48a060dc7347be84ea9f52d9f",
-                "AAAAA1c8ELFI/EvH25dUDOSdoikw9LzUigYNxzR76E6p9S2f",
-            ),
-        ],
-    )
-    def test_hash_memo(self, hex, xdr):
+
+    def test_hash_memo(self):
+        hex = "573c10b148fc4bc7db97540ce49da22930f4bcd48a060dc7347be84ea9f52d9f"
+        xdr = "AAAAA1c8ELFI/EvH25dUDOSdoikw9LzUigYNxzR76E6p9S2f"
         hash = binascii.unhexlify(hex)
         memo = HashMemo(hash)
-        assert memo.to_xdr_object().to_xdr() == xdr
+        assert memo.to_xdr() == xdr
+        assert Memo.from_xdr(xdr) == memo
 
         base_memo = Memo.from_xdr_object(memo.to_xdr_object())
         assert isinstance(base_memo, HashMemo)
-        assert base_memo.to_xdr_object().to_xdr() == xdr
+        assert base_memo.to_xdr() == xdr
 
     def test_hash_memo_too_long_raise(self):
         length = 33
         with pytest.raises(
             MemoInvalidException,
-            match="HashMemo can contain 32 bytes at max, got {:d} bytes".format(length),
+            match="The length of HashMemo should be 32 bytes, got {:d} bytes.".format(length),
         ):
             HashMemo(os.urandom(length))
 
-    @pytest.mark.parametrize(
-        "hex, xdr",
-        [
-            (
-                "90d38f2d64949bb64f693ff231f9a3b3",
-                "AAAABJDTjy1klJu2T2k/8jH5o7MAAAAAAAAAAAAAAAAAAAAA",
-            ),
-            (
-                "573c10b148fc4bc7db97540ce49da22930f4bcd48a060dc7347be84ea9f52d9f",
-                "AAAABFc8ELFI/EvH25dUDOSdoikw9LzUigYNxzR76E6p9S2f",
-            ),
-        ],
-    )
-    def test_return_hash_memo(self, hex, xdr):
+    def test_hash_memo_too_short_raise(self):
+        length = 16
+        with pytest.raises(
+            MemoInvalidException,
+            match="The length of HashMemo should be 32 bytes, got {:d} bytes.".format(length),
+        ):
+            HashMemo(os.urandom(length))
+
+    def test_return_hash_memo(self):
+        hex = "573c10b148fc4bc7db97540ce49da22930f4bcd48a060dc7347be84ea9f52d9f"
+        xdr = "AAAABFc8ELFI/EvH25dUDOSdoikw9LzUigYNxzR76E6p9S2f"
         return_hash = binascii.unhexlify(hex)
         memo = ReturnHashMemo(return_hash)
-        assert memo.to_xdr_object().to_xdr() == xdr
+        assert memo.to_xdr() == xdr
+        assert Memo.from_xdr(xdr) == memo
 
         base_memo = Memo.from_xdr_object(memo.to_xdr_object())
         assert isinstance(base_memo, ReturnHashMemo)
-        assert base_memo.to_xdr_object().to_xdr() == xdr
+        assert base_memo.to_xdr() == xdr
 
     def test_return_hash_memo_too_long_raise(self):
-        length = 33
+        length = 48
         with pytest.raises(
             MemoInvalidException,
-            match="ReturnHashMemo can contain 32 bytes at max, got {:d} bytes".format(
+            match="The length of ReturnHashMemo should be 32 bytes, got {:d} bytes.".format(
+                length
+            ),
+        ):
+            ReturnHashMemo(os.urandom(length))
+
+    def test_return_hash_memo_too_short_raise(self):
+        length = 16
+        with pytest.raises(
+            MemoInvalidException,
+            match="The length of ReturnHashMemo should be 32 bytes, got {:d} bytes.".format(
                 length
             ),
         ):